requirement.txt
pytorch>1.0.0
tnt
fire
tqdm
numpy
tensorboardX